+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # DuckyLib Changelog
 
+## 8/7/25
+
+- added default vertex and fragment shaders to `constants.h`, if `NULL` is passed in creating a new entity it swaps it for those instead
+- removed `DL_FRAG_SHADER_DEFAULT` and `DL_VERT_SHADER_DEFAULT` from `renderer.h`
+
 ## 6/7/25
 
 - fixed the dates in `CHANGELOG.md`